fix: move new template methods to PRService
