+++ apps/desktop/src/lib/gitHost/github/githubPrService.ts
@@ -1,16 +1,13 @@
 import { GitHubPrMonitor } from './githubPrMonitor';
 import { DEFAULT_HEADERS } from './headers';
 import { ghResponseToInstance, parseGitHubDetailedPullRequest } from './types';
+import { invoke } from '$lib/backend/ipc';
+import { showToast } from '$lib/notifications/toasts';
 import { sleep } from '$lib/utils/sleep';
 import posthog from 'posthog-js';
 import { writable } from 'svelte/store';
 import type { GitHostPrService } from '$lib/gitHost/interface/gitHostPrService';
-import type {
-	CreatePullRequestArguments,
-	DetailedPullRequest,
-	MergeMethod,
-	PullRequest
-} from '$lib/gitHost/interface/types';
+import type { DetailedPullRequest, MergeMethod, PullRequest } from '$lib/gitHost/interface/types';
 import type { RepoInfo } from '$lib/url/gitUrl';
 import type { Octokit } from '@octokit/rest';
 
@@ -85,5 +82,34 @@ export class GitHubPrService implements GitHostPrService {
 
 	prMonitor(prNumber: number): GitHubPrMonitor {
 		return new GitHubPrMonitor(this, prNumber);
+	}
+
+	async getPrTemplateContent(path: string) {
+		try {
+			const fileContents: string | undefined = await invoke('get_pr_template_contents', { path });
+			return fileContents;
+		} catch (err) {
+			console.error(`Error reading pull request template at path: ${path}`, err);
+
+			showToast({
+				title: 'Failed to read pull request template',
+				message: `Could not read: \`${path}\`.`,
+				style: 'neutral'
+			});
+		}
+	}
+
+	async getAvailablePrTemplates(path: string): Promise<string[] | undefined> {
+		// TODO: Find a workaround to avoid this dynamic import
+		// https://github.com/sveltejs/kit/issues/905
+		const { join } = await import('@tauri-apps/api/path');
+		const targetPath = await join(path, '.github');
+
+		const availableTemplates: string[] | undefined = await invoke(
+			'available_pull_request_templates',
+			{ path: targetPath }
+		);
+
+		return availableTemplates;
 	}
 }
